+++ docs/docs/changelog.md
@@ -3,6 +3,7 @@
 - Added support for icons in the page title 
 - Fixed issue where todo widget highlight was too short
 - Fixed issue where in todo widget trash animation icon was slower than highlight
+- Added speedtest widget
 
 # Changes for 2.3.1
 - Added support for loading environment variables from a file via `--env-file`